6.18.2024

+++ farm_cmd/ez_farm.py
@@ -1,9 +1,13 @@
 import random
 import time
 import subprocess as sp
+import logging
+import psutil
+
+logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
 start_delay = 10
-close_delay = 10
+close_delay = 90
 
 processes = [
     {"game_id": "2923300", "title": "Banana", "name": "Banana.exe", "start_delay": start_delay, "close_delay": close_delay},
@@ -13,33 +17,45 @@ processes = [
     {"game_id": "2996990", "title": "Flaggenspiel", "name": "Flaggenspiel.exe", "start_delay": start_delay, "close_delay": close_delay}
 ]
 
+def is_process_running(process_name):
+    for proc in psutil.process_iter(['name']):
+        if proc.info['name'] == process_name:
+            return True
+    return False
+
 def run_process(game_id, window_title, process_name, delay_click, delay_close):
-    process = sp.Popen(f'start steam://rungameid/{game_id}', shell=True)
-    time.sleep(delay_click)
-    print(f"Game {window_title} started. Waiting for {delay_close} seconds before closing...")
-    time.sleep(delay_close)
-    print(f"Closing game {window_title}...")
-    sp.Popen(f'taskkill /IM "{process_name}" /F', shell=True)
+    try:
+        process = sp.Popen(f'start steam://rungameid/{game_id}', shell=True)
+        logging.info(f"✨ Game {window_title} started. Waiting for {delay_click} seconds to check if it is running...")
+        time.sleep(delay_click)
+
+        if is_process_running(process_name):
+            logging.info(f"✅ Game {window_title} is running. Waiting for {delay_close} seconds before closing...")
+            time.sleep(delay_close)
+            logging.info(f"✖️ Closing game {window_title}...")
+            sp.Popen(f'taskkill /IM "{process_name}" /F', shell=True)
+        else:
+            logging.warning(f"⚠️ Game {window_title} did not start. Skipping to the next game...")
+    except Exception as e:
+        logging.error(f"❌ An error occurred: {e}")
 
 def start_farming():
-    global farming
-    farming = True
     pint = 0
-    while farming:
-        print(f'Start loop: {pint}...')
-
-        pint += 1
+    while True:
+        logging.info(f"🚩 Starting loop: {pint}")
         for process in processes:
-            if not farming:
-                break
-            print(f"Opening game: '{process["title"]}'...")
+            logging.info(f"🎮 Opening game: '{process['title']}'...")
             run_process(process["game_id"], process["title"], process["name"], process["start_delay"], process["close_delay"])
 
-        #new_loop_start_time = random.randint(5, 15)
-        new_loop_start_time = random.randint(10800, 12600)  # 3 - 3.5h
-        last_pint = pint - 1
-
-        print(f'Loop {last_pint} will finish in {new_loop_start_time}... sec')
+        new_loop_start_time = random.randint(10800, 11000)  # 3h+-
+        last_pint = pint
+        pint += 1
+        logging.info(f"✅ Loop: {last_pint} finished!")
+        logging.info(f"🕒 Waiting for {new_loop_start_time} seconds before starting the next loop: {pint}...")
         time.sleep(new_loop_start_time)
 
-start_farming()
+if __name__ == "__main__":
+    try:
+        start_farming()
+    except KeyboardInterrupt:
+        logging.info("⛔ Stopping farming...")
